hw/intc/arm_gicv3_its_common: Increase DeviceID to 32 bits
According to the “GICv3 and GICv4 Software Overview” the DeviceID is
IMPLEMENTATION DEFINED. This patch increases the DeviceID in send_msi virtual
function to 32 bits to allow the possibility of a redefinition of it with a
bigger DeviceID.

Message-Id: <20240221173325.1494895-2-nabihestefan@google.com>

+++ hw/intc/arm_gicv3_its_kvm.c
@@ -42,7 +42,7 @@ struct KVMARMITSClass {
 };
 
 
-static int kvm_its_send_msi(GICv3ITSState *s, uint32_t value, uint16_t devid)
+static int kvm_its_send_msi(GICv3ITSState *s, uint32_t value, uint32_t devid)
 {
     struct kvm_msi msi;
